requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
google-cloud-aiplatform==1.71.1
google-generativeai==0.3.1
python-dotenv==1.0.0
typing-extensions==4.8.0
vertexai==1.71.1
